+++ packages/base/core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Don't load developer tools automatically in production
+
 ## [1.1.0] - 2025-01-10
 
 ### Added